chore: unify all user agents
