+++ packages/playwright-core/src/utils/browserFetcher.ts
@@ -20,9 +20,8 @@ import fs from 'fs';
 import os from 'os';
 import path from 'path';
 import { execSync } from 'child_process';
-import { existsAsync, download, getPlaywrightVersion } from './utils';
+import { existsAsync, download, getPlaywrightVersion, getUserAgent } from './utils';
 import { debugLogger } from './debugLogger';
-import { parseOSReleaseText } from './ubuntuVersion';
 
 export async function downloadBrowserWithProgressBar(title: string, browserDirectory: string, executablePath: string, downloadURL: string, downloadFileName: string): Promise<boolean> {
   const progressBarName = `Playwright build of ${title}`;
@@ -38,7 +37,7 @@ export async function downloadBrowserWithProgressBar(title: string, browserDirec
     await download(url, zipPath, {
       progressBarName,
       log: debugLogger.log.bind(debugLogger, 'install'),
-      userAgent: getDownloadUserAgent(),
+      userAgent: getUserAgent(),
     });
     debugLogger.log('install', `extracting archive`);
     debugLogger.log('install', `-- zip: ${zipPath}`);
@@ -65,50 +64,4 @@ export function logPolitely(toBeLogged: string) {
 
   if (!logLevelDisplay)
     console.log(toBeLogged);  // eslint-disable-line no-console
-}
-
-let cachedUserAgent: string | undefined;
-function getDownloadUserAgent(): string {
-  if (cachedUserAgent)
-    return cachedUserAgent;
-  try {
-    cachedUserAgent = determineDownloadUserAgent();
-  } catch (e) {
-    cachedUserAgent = 'Playwright/unknown';
-  }
-  return cachedUserAgent;
-}
-
-function determineDownloadUserAgent(): string {
-  let osIdentifier = 'unknown';
-  let osVersion = 'unknown';
-  if (process.platform === 'win32') {
-    const version = os.release().split('.');
-    osIdentifier = 'windows';
-    osVersion = `${version[0]}.${version[1]}`;
-  } else if (process.platform === 'darwin') {
-    const version = execSync('sw_vers -productVersion').toString().trim().split('.');
-    osIdentifier = 'macOS';
-    osVersion = `${version[0]}.${version[1]}`;
-  } else if (process.platform === 'linux') {
-    try {
-      const osReleaseText = fs.readFileSync('/etc/os-release', 'utf8');
-      const fields = parseOSReleaseText(osReleaseText);
-      osIdentifier = fields.get('id') || 'unknown';
-      osVersion = fields.get('version_id') || 'unknown';
-    } catch (e) {
-    }
-  }
-
-  let langName = 'unknown';
-  let langVersion = 'unknown';
-  if (!process.env.PW_CLI_TARGET_LANG) {
-    langName = 'node';
-    langVersion = process.version.substring(1).split('.').slice(0, 2).join('.');
-  } else if (['node', 'python', 'java', 'csharp'].includes(process.env.PW_CLI_TARGET_LANG)) {
-    langName = process.env.PW_CLI_TARGET_LANG;
-    langVersion = process.env.PW_CLI_TARGET_LANG_VERSION ?? 'unknown';
-  }
-
-  return `Playwright/${getPlaywrightVersion()} (${os.arch()}; ${osIdentifier} ${osVersion}) ${langName}/${langVersion}`;
 }
